refactor(variant): isolate implementation details
