+++ include/cbor_tags/cbor_concepts.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "cbor_tags/cbor_variant_traits.h"
 #include "cbor_tags/detail/cbor_range_concepts.h"
+#include "cbor_tags/detail/cbor_variant_traits.h"
 
 #include <array>
 #include <concepts>
